power: qpnp-charger: Restore charge current to max after thermal throttle
When thermal throttling is over and we're back to a normal operating temperature, the charge current is not restored to what it was before throttling (max_bat_chg_current).

Set the charge current back to the max instead of 1200mA when thermal mitigation is finished in order to fix the slow-charge issue.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

Conflicts:
	drivers/power/qpnp-charger.c